--- TodoApp-BD/README.md
@@ -1,2 +1,0 @@
-# TodoApp
-Gerenciador de Tarefas - UC POO
