--- TodoApp-BD/resources/comandos_mysql.sql
@@ -1,3 +1,0 @@
-CREATE DATABASE todoapp;
-USE todoapp;
-CREATE TABLE tarefa (id int PRIMARY KEY AUTO_INCREMENT, nome varchar(255), descricao varchar(255));
